+++ src/app/component/MissionTypeModal/index.js
@@ -1,0 +1,3 @@
+import MissionTypeModal from "./MissionTypeModal";
+
+export default MissionTypeModal;
